refactoring SearchContext
