+++ frontend/src/components/SearchContext/index.ts
@@ -1,0 +1,3 @@
+// eslint-disable-next-line no-restricted-exports
+export { default } from './SearchContext';
+export { useSearchContext } from './hooks';
