+++ frontend/src/components/SearchSubmit/SearchSubmit.tsx
@@ -1,16 +1,27 @@
 import { createSearchParams, useNavigate } from 'react-router-dom';
+import { useCallback } from 'react';
 import { addHistory } from '../../store';
 import { SEARCH } from '../../utils/const';
 import { useAppDispatch } from '../../hooks/redux';
-import { useSearchContext } from '../Search/hooks';
+import { useSearchContext } from '../SearchContext';
 import UseGetSearchParams from '../../hooks/useGetSearchParams';
+import useOnClickOutside from '../../hooks/useOnClickOutside';
+import { Props } from './type';
 
-function SearchSubmit() {
+function SearchSubmit(props: Props) {
+  const { searchRef } = props;
   const { limit } = UseGetSearchParams();
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
 
-  const { search, type } = useSearchContext();
+  const { search, type, setResults } = useSearchContext();
+  const auth = useGetAuth();
+
+  const suggestClose = useCallback(() => {
+    setResults([]);
+  }, []);
+
+  useOnClickOutside(searchRef, suggestClose);
 
   const submitHandler = () => {
     dispatch(addHistory({
