+++ frontend/src/components/SearchSubmit/type.ts
@@ -1,0 +1,5 @@
+import { RefObject } from 'react';
+
+export type Props = {
+  searchRef: RefObject<HTMLDivElement>
+};
